Consolidate tests and reorder checks to avoid extra queries
(Changes following review comments)
