+++ spec/system/candidate_interface/candidate_reviewing_application_with_unavailable_course_options_spec.rb
@@ -15,6 +15,14 @@ RSpec.feature 'Candidate reviewing an application with unavailable course option
 
     when_i_submit_the_application
     then_i_see_error_messages_for_the_things_i_was_warned_about
+
+    when_i_swap_course_option_with_one_unavailable_on_apply
+
+    when_i_visit_the_review_application_page
+    then_i_see_a_warning_that_my_course_is_no_longer_on_apply
+
+    when_i_submit_the_application
+    then_i_see_error_messages_for_the_course_closed_on_apply_i_was_warned_about
   end
 
   def given_i_am_signed_in
@@ -85,6 +93,24 @@ RSpec.feature 'Candidate reviewing an application with unavailable course option
     end
   end
 
+  def when_i_swap_course_option_with_one_unavailable_on_apply
+    @option_where_course_not_running.course.update!(
+      exposed_in_find: true,
+      open_on_apply: false,
+    )
+  end
+
+  def then_i_see_a_warning_that_my_course_is_no_longer_on_apply
+    expect(page).to have_content(course_closed_on_apply_message)
+    expect(page).to have_content("You can still apply for '#{@option_where_course_not_running.course.name_and_code}' on UCAS.")
+  end
+
+  def then_i_see_error_messages_for_the_course_closed_on_apply_i_was_warned_about
+    within('.govuk-error-summary') do
+      expect(page).to have_content course_closed_on_apply_message
+    end
+  end
+
 private
 
   def course_not_running_message
@@ -97,5 +123,9 @@ private
 
   def chosen_site_has_no_vacancies_message
     "Your chosen site for '#{@option_where_no_vacancies_at_chosen_site.course.provider_and_name_code}' has no vacancies"
+  end
+
+  def course_closed_on_apply_message
+    "'#{@option_where_course_not_running.course.name_and_code}' at #{@option_where_course_not_running.course.provider.name} is not available on Apply for teacher training anymore"
   end
 end
